+++ apps/annotations/graphql/types.py
@@ -1,0 +1,5 @@
+from ariadne import ObjectType, QueryType, MutationType
+
+annotation = ObjectType("Annotation")
+query = QueryType()
+mutation = MutationType()