+++ apps/audit/apps.py
@@ -18,7 +18,8 @@ class AuditConfig(AppConfig):
         from apps.invoicing import models as invoicing
         from apps.stats import models as stats
         from apps.wiki import models as wiki
-        from apps.employees import models as empployees
+        from apps.employees import models as employees
+        from apps.annotations import models as annotations
 
         models = [
             warehouse.Ware,
@@ -42,8 +43,9 @@ class AuditConfig(AppConfig):
             wiki.ExternalLink,
             wiki.ArticleFile,
             wiki.Tag,
-            empployees.Employee,
-            empployees.WorkAbsence,
+            employees.Employee,
+            employees.WorkAbsence,
+            annotations.Annotation,
         ]
 
         for model in models: