+++ apps/invoicing/apps.py
@@ -2,4 +2,10 @@ from django.apps import AppConfig
 
 
 class InvoicingConfig(AppConfig):
-    name = "invoicing"
+    name = "apps.invoicing"
+
+    def ready(self):
+        from apps.annotations.registry import annotations
+        from apps.invoicing.models import Contractor
+
+        annotations.register(Contractor)